Update the README
This change adds the missing supporting image in the README file, and
extends it to be more complete. Specifically, it adds the standard
ending sections (contributing, did you find the project useful, etc). It
also reworks the existing content a little to be more engaging, concise,
and specific, and moves all URLs to be reference-style links.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>